Fix top-level tests
